requirements.txt
fastapi
uvicorn
langchain
langchain-openai
openai
python-dotenv
redis
langsmith
langchainhub
sqlalchemy[asyncio]
databases
alembic
aiomysql
pymysql
cryptography
langchain_community